+++ src/Adapters/SpaceEngineers2/Kontrol.Adapters.SpaceEngineers2.Tests/CockpitInputPatchTests.cs
@@ -581,7 +581,7 @@ public class CockpitInputPatchTests
     {
         var schema = new SpaceEngineers2Installer().GetInputSchema();
 
-        schema.Version.ShouldBe(8);
+        schema.Version.ShouldBe(9);
         schema.Inputs[10].Id.ShouldBe("systems.exit_grid");
         schema.Inputs[6].Id.ShouldBe("systems.dampeners");
         schema.Inputs[6].DiscreteBehavior.ShouldBe(DiscreteBehavior.Toggle);
@@ -607,6 +607,34 @@ public class CockpitInputPatchTests
         schema.Inputs[15].EffectiveDeliveryMode.ShouldBe(DiscreteDeliveryMode.State);
         schema.Inputs[16].EffectiveActionBehavior.ShouldBe(DiscreteBehavior.Momentary);
         schema.Inputs[16].EffectiveDeliveryMode.ShouldBe(DiscreteDeliveryMode.State);
+
+        for (int i = 0; i < BeltSelectionPatch.ActionCount; i++)
+        {
+            var input = schema.Inputs[17 + i];
+            string key = i == 9 ? "0" : (i + 1).ToString();
+            input.Id.ShouldBe($"belt.select_{key}");
+            input.Category.ShouldBe("Belt");
+            input.SignalKind.ShouldBe(InputSignalKind.Discrete);
+            input.EffectiveActionBehavior.ShouldBe(DiscreteBehavior.Trigger);
+            input.EffectiveDeliveryMode.ShouldBe(DiscreteDeliveryMode.Event);
+        }
+    }
+
+    [TestCase(17, 0)]
+    [TestCase(18, 1)]
+    [TestCase(25, 8)]
+    [TestCase(26, 9)]
+    public void BeltActionBits_MapToTheExpectedToolbarTile(int actionBit, int expectedTileIndex)
+    {
+        BeltSelectionPatch.TryGetTileIndex(actionBit, out int tileIndex).ShouldBeTrue();
+        tileIndex.ShouldBe(expectedTileIndex);
+    }
+
+    [TestCase(16)]
+    [TestCase(27)]
+    public void BeltActionBits_RejectBitsOutsideTheTenSlotRange(int actionBit)
+    {
+        BeltSelectionPatch.TryGetTileIndex(actionBit, out _).ShouldBeFalse();
     }
 
     [Test]